+++ src/test/kotlin/com/wutsi/codegen/kotlin/server/ServerCLITest.kt
@@ -45,12 +45,12 @@ internal class ServerCLITest {
 
         val args = arrayOf(
             "server",
-            "-a like",
-            "-p com.wutsi.test",
-            "-i $url",
-            "-o ./target",
-            "-j 11",
-            "-g foo"
+            "-name", "like",
+            "-package", "com.wutsi.test",
+            "-in", "$url",
+            "-out", "./target",
+            "-jdk", "11",
+            "-github_user", "foo"
         )
         cli.run(args)
 
@@ -75,10 +75,10 @@ internal class ServerCLITest {
 
         val args = arrayOf(
             "server",
-            "-a like",
-            "-p com.wutsi.test",
-            "-i $url",
-            "-o ./target"
+            "-name", "like",
+            "-package", "com.wutsi.test",
+            "-in", "$url",
+            "-out", "./target",
         )
         cli.run(args)