+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # See https://keepachangelog.com/en/1.0.0/
 
+## [0.0.20] 2021-03-31
+### Added
+- README generator
+- Add swagger-ui code generator
+
+### Changed
+- Do not overide `application..yml` files
+- Add missing repository `wutsi-stream-rabbitmq` in `pom.xml`
+- Exclude from code coverage configuration classes
+
 ## [0.0.19] 2021-03-29
 ### Changed
 - Integration of Message Queue
 - Change caching configuration file from `StringCache...` to `Cache...`
-### Fixes
+### Fixed
 - Update dependency of `wutsi-stream-memcached` to `v1.1.0` to fix a NPE
 - FIX: RabbitMQ configuration variables (`rabbitnq...` -> `rabbitmq...`)